requirements.txt
langchain>=0.2.11
langchain-groq>=0.1.5
langchain-huggingface>=0.0.8
langchain-community>=0.2.10
langchain-chroma>=0.1.2
chromadb>=0.5.3
pydantic>=2.7.4
python-dotenv>=1.0.1
rich>=13.7.1
groq>=0.9.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
